Don't fail if build exists and also don't remove build dir if it already exists

+++ lib/galaxy/tools/deps/mulled/mulled_build.py
@@ -275,12 +275,16 @@ class InvolucroContext(installable.InstallableContext):
     def exec_command(self, involucro_args):
         cmd = self.build_command(involucro_args)
         # Create ./build dir manually, otherwise Docker will do it as root
-        os.mkdir('./build')
+        created_build_dir = False
+        if not os.path.exists('build'):
+            created_build_dir = True
+            os.mkdir('./build')
         try:
             res = self.shell_exec(" ".join(cmd))
         finally:
             # delete build directory in any case
-            shutil.rmtree('./build')
+            if created_build_dir:
+                shutil.rmtree('./build')
         return res
 
     def is_installed(self):
